\name{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@}
\title{Bioconductor annotation data package}
\description{
  
  Welcome to the @ANNOBJPREFIX@.db annotation Package.  The purpose of
  this package is to provide detailed information about the
  @ANNOBJPREFIX@ platform. This package is updated biannually.
  
  This annotation package supports the following objects:

  ls("package:@ANNOBJPREFIX@.db") 
  
  Each of these objects has their own manual page detailing where
  relevant data was obtained along with some examples of how to use it.

\keyword{datasets}

